% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chads2Model.R
\name{chads2Model}
\alias{chads2Model}
\title{Apply the existing model CHADs2 using the standardised framework}
\usage{
chads2Model(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  outcomeDatabaseSchema, cohortTable, outcomeTable, cohortId, outcomeId,
  oracleTempSchema = NULL, riskWindowStart = 1, riskWindowEnd = 365,
  requireTimeAtRisk = T, minTimeAtRisk = 364, includeAllOutcomes = T,
  firstExposureOnly = F, removePriorOutcome = T)
}
\arguments{
\item{connectionDetails}{The connection details for extracting the data}

\item{cdmDatabaseSchema}{A string specifying the database containing the cdm}

\item{cohortDatabaseSchema}{A string specifying the database containing the target population cohort}

\item{outcomeDatabaseSchema}{A string specifying the database containing the outcome cohort}

\item{cohortTable}{A string specifying the table containing the target population cohort}

\item{outcomeTable}{A string specifying the table containing the outcome cohort}

\item{cohortId}{An iteger specifying the cohort id for the target population cohorts}

\item{outcomeId}{An iteger specifying the cohort id for the outcome cohorts}

\item{oracleTempSchema}{The temp schema require is using oracle}

\item{riskWindowStart}{The start of the period to predict the risk of the outcome occurring start relative to the target cohort start date}

\item{riskWindowEnd}{The end of the period to predict the risk of the outcome occurring start relative to the target cohort start date}

\item{requireTimeAtRisk}{Require a minimum number of days observed in the time at risk period?}

\item{minTimeAtRisk}{If requireTimeAtRisk is true, the minimum number of days at risk}

\item{includeAllOutcomes}{Whether to include people with outcome who do not satify the minTimeAtRisk}

\item{firstExposureOnly}{Whether to restrict to first target cohort start date if people are in target popualtion multiple times}

\item{removePriorOutcome}{Remove people with prior outcomes from the target population}
}
\value{
A list containing the model performance and the personal predictions for each subject in the target population
}
\description{
Apply the existing model CHADs2 using the standardised framework
}
\details{
This function applies CHADs2 to a target cohort and validates the performance
}
