% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{extractNetBenefit}
\alias{extractNetBenefit}
\title{The net benefit for a model}
\usage{
extractNetBenefit(runPlp, type = NULL, modelId = NULL)
}
\arguments{
\item{runPlp}{A object of class 'runPlp' returned by implementing runPlp() or the output when implementing an existing model}
}
\value{
A data.frame containing the columns: pt, netbenefit, treatall and modelId.  Pt is the probability treshold or a scaled risk score (if using an existing model)
}
\description{
This function calculates the net benefit for a model
}
\details{
Users need to input a trained model (the output of runPlp()) or the output of running an existing model
}
