% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{getThresholds}
\alias{getThresholds}
\title{Get prediction risk thresholds for quantiles}
\usage{
getThresholds(plpModel1, plpModel2, percentage = c(0.99, 0.95, 0.9, 0.5))
}
\arguments{
\item{plpModel1}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}

\item{plpModel2}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}

\item{percentage}{A quantile percentage (e.g., c(0.99, 0.95)) will find the threshold that only 1 percent and 5 percent of people have a risk equal to or higher than for both models}
}
\value{
A data.frame with each column being the risk thresholds per model and rows corresponding to the input quantiles
}
\description{
This function calculates the net reclassification improvement (NRI) metric comparing two models at a range of predicting tresholds
}
\details{
Users need to input a trained model (the output of runPlp()) or the output of running an existing model
}
