% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{IDI}
\alias{IDI}
\title{integrated discrimination improvement- no thresholds needed (an approximation)}
\usage{
IDI(plpModel1, plpModel2)
}
\arguments{
\item{plpModel1}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}

\item{plpModel2}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}
}
\value{
A list containing the IDI value and z-value (can be used for statistical significance). A positive value suggests the first model is better than the second model.  A negative values suggests the opposite.
}
\description{
This function calculates the ntegrated discrimination improvement (IDI) metric comparing two models
}
\details{
Users need to input two trained models (the output of runPlp()) or the output of running an existing model
}
