% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelNames-data.R
\docType{data}
\name{modelNames}
\alias{modelNames}
\title{The names for each model (with the corresponding modelId)}
\format{A csv table with the columns: modelId, Name}
\usage{
data(modelNames)
}
\description{
This table contains the existing model names in the package
}
\examples{
data(modelNames)
head(modelNames)
}
\keyword{datasets}
