% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{multipleIDI}
\alias{multipleIDI}
\title{multiple integrated discrimination improvement}
\usage{
multipleIDI(plpRunList)
}
\arguments{
\item{plpRunList}{A list of objects returned  runPlp() containing the trained model or the output when implementing an existing model}
}
\value{
A data.frame containing the columns: Model1, Model2, IDI value and z-value (can be used for statistical significance). A positive value suggests Model1 is better than Model2.  A negative values suggests the opposite.
}
\description{
This function calculates the ntegrated discrimination improvement (IDI) metric for all model comparisons
}
\details{
Users need to input a lsit of two or more trained models (the output of runPlp()) or the output of running an existing model
}
