% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NRI2}
\alias{NRI2}
\title{Net reclassification improvement 2 no thresholds}
\usage{
NRI2(plpModel1, plpModel2)
}
\arguments{
\item{plpModel1}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}

\item{plpModel2}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}
}
\value{
A list containing the NRI value and z-value (can be used for statistical significance).
A positive value suggests the first model is better than the second model.  A negative values suggests the opposite.
}
\description{
This function calculates a modified net reclassification improvement (NRI) metric comparing two models at a range of predicting tresholds (simialr to library(Hmisc))
}
\details{
Users need to input a trained model (the output of runPlp()) or the output of running an existing model
}
