% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDecisionCurve}
\alias{plotDecisionCurve}
\title{Plot a decsion curve for multiple models}
\usage{
plotDecisionCurve(runPlpList, names = NULL, type = NULL)
}
\arguments{
\item{runPlpList}{An object of class 'runPlp' or list of object of class 'runPlp' returned by implementing runPlp() or the output when implementing an existing model}

\item{names}{A string or vector of model names corresponding to the runPlpList}

\item{type}{A strng or vector of evaluation types (train/test/validation) to filter for each runPlp}
}
\value{
A plot is returned with the model decision cruves
}
\description{
This function plots a decision curve
}
\details{
Users need to input a list of trained models (the output of runPlp()) or the output of running an existing model
}
