% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotModelStratification2}
\alias{plotModelStratification2}
\title{Plot a the model performance visulisation stratified by risk}
\usage{
plotModelStratification2(runPlp, specificity = 0.95)
}
\arguments{
\item{runPlp}{An object of class 'runPlp' or list of object of class 'runPlp' returned by implementing runPlp() or the output when implementing an existing model}

\item{specificity}{The required specificity value for the model}
}
\value{
A plot is returned with the overal outcome risk and observed risk in different model risk groups
}
\description{
This function plots a visulisation of model performance stratified by risk
}
\details{
Users need to input a trained model (the output of runPlp()) or the output of running an existing model
}
