% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotModelStratification3}
\alias{plotModelStratification3}
\title{Plot a the model performance visulisation stratified by risk}
\usage{
plotModelStratification3(runPlp, type = "test", riskGroups = NULL)
}
\arguments{
\item{runPlp}{An object of class 'runPlp' or list of object of class 'runPlp' returned by implementing runPlp() or the output when implementing an existing model}

\item{type}{test/train/validation}

\item{riskGroups}{...}
}
\value{
A plot is returned with the overal outcome risk and observed risk in different model risk groups
}
\description{
This function plots a visulisation of model performance stratified by risk
}
\details{
Users need to input a trained model (the output of runPlp()) or the output of running an existing model
}
