% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMultipleCal}
\alias{plotMultipleCal}
\title{Plot a calibration plot for multiple models}
\usage{
plotMultipleCal(runPlpList, names, type = "test", grid = F, ncol = 2)
}
\arguments{
\item{runPlpList}{An object of class 'runPlp' or list of object of class 'runPlp' returned by implementing runPlp() or the output when implementing an existing model}

\item{names}{A string or vector of model names corresponding to the runPlpList}

\item{type}{A strng or vector of evaluation types (train/test/validation) to filter for each runPlp}

\item{grid}{Whether to plot each model seperately using a grid or overlay the plots (grid=F)}

\item{ncol}{The number of columns for the grid}
}
\value{
A plot is returned with the calibration plots for each model
}
\description{
This function plots multiple calibration plots for comparison
}
\details{
Users need to input a list of trained models (the output of runPlp()) or the output of running an existing model
}
