% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NRI}
\alias{NRI}
\title{Net reclassification improvement}
\usage{
NRI(plpModel1, plpModel2, thresholds = seq(0, 1, 1/100),
  secondThresholds = NULL)
}
\arguments{
\item{plpModel1}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}

\item{plpModel2}{The object returned by runPlp() containing the trained model or the output when implementing an existing model}

\item{thresholds}{A sequence of predicting tresholds to calcuate the NRI at each of these tresholds}

\item{secondThresholds}{A sequence of predicting tresholds to calcuate the NRI at each of these tresholds for the second model (if NULL uses thresholds)}
}
\value{
A list containing the NRI value and z-value (can be used for statistical significance) plus various values used to calculate the
NRI for each thresold. A positive value suggests the first model is better than the second model.  A negative values suggests the opposite.
}
\description{
This function calculates the net reclassification improvement (NRI) metric comparing two models at a range of predicting tresholds
}
\details{
Users need to input a trained model (the output of runPlp()) or the output of running an existing model
}
