% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstrokeModel.R
\name{qstrokeModel}
\alias{qstrokeModel}
\title{Apply the existing model Qstroke stroke risk using the standardised framework}
\usage{
qstrokeModel(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  outcomeDatabaseSchema,
  cohortTable,
  outcomeTable,
  cohortId,
  outcomeId,
  oracleTempSchema = NULL,
  cdmVersion = 5,
  analysisId = "Q-stroke-female",
  sampleSize = NULL,
  riskWindowStart = 1,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  endDay = -1,
  requireTimeAtRisk = T,
  minTimeAtRisk = 1,
  includeAllOutcomes = T,
  removePriorOutcome = T,
  priorOutcomeLookback = 9999,
  calibrationPopulation = NULL,
  createCovariateCohorts = T,
  overwriteExistingCohorts = T,
  covariateDatabaseSchema = cohortDatabaseSchema,
  covariateCohortTable = "PredictionComp"
)
}
\arguments{
\item{connectionDetails}{The connection details for extracting the data}

\item{cdmDatabaseSchema}{A string specifying the database containing the cdm}

\item{cohortDatabaseSchema}{A string specifying the database containing the target population cohort}

\item{outcomeDatabaseSchema}{A string specifying the database containing the outcome cohort}

\item{cohortTable}{A string specifying the table containing the target population cohort}

\item{outcomeTable}{A string specifying the table containing the outcome cohort}

\item{cohortId}{An iteger specifying the cohort id for the target population cohorts}

\item{outcomeId}{An iteger specifying the cohort id for the outcome cohorts}

\item{oracleTempSchema}{The temp schema require is using oracle}

\item{cdmVersion}{The CDM version you are using}

\item{analysisId}{An id to give to the study}

\item{sampleSize}{The size to sample from T when applying the existing model}

\item{riskWindowStart}{The start of the period to predict the risk of the outcome occurring start relative to the target cohort start date}

\item{startAnchor}{Is the risk start relative to cohort_start or cohort_end}

\item{riskWindowEnd}{The end of the period to predict the risk of the outcome occurring start relative to the target cohort start date}

\item{endAnchor}{Is the risk end relative to cohort_start or cohort_end}

\item{endDay}{The last day relative to index for the covariates}

\item{requireTimeAtRisk}{Require a minimum number of days observed in the time at risk period?}

\item{minTimeAtRisk}{If requireTimeAtRisk is true, the minimum number of days at risk}

\item{includeAllOutcomes}{Whether to include people with outcome who do not satify the minTimeAtRisk}

\item{removePriorOutcome}{Remove people with prior outcomes from the target population}

\item{priorOutcomeLookback}{Time prior to index to remove people if they have the outcome}

\item{calibrationPopulation}{A data.frame of subjectId, cohortStartDate, indexes used to recalibrate the model on new data}

\item{createCovariateCohorts}{Create the cohorts used for the covariates in the model}

\item{overwriteExistingCohorts}{Overwrite cohorts that already exist in the covariateCohortTable table}

\item{covariateDatabaseSchema}{A read/write database to keep the covariate cohorts}

\item{covariateCohortTable}{A read/write table to keep the covariate cohorts}
}
\value{
A list containing the model performance and the personal predictions for each subject in the target population
}
\description{
Apply the existing model Qstroke stroke risk using the standardised framework
}
\details{
This function applies Qstroke stroke risk to a target cohort and validates the performance
}
