% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstrokeModel.R
\name{qstrokeModel}
\alias{qstrokeModel}
\title{Apply the existing model Qstroke stroke risk using the standardised framework}
\usage{
qstrokeModel(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  outcomeDatabaseSchema, cohortTable, outcomeTable, cohortId, outcomeId,
  oracleTempSchema = NULL, riskWindowStart = 1, riskWindowEnd = 365,
  addExposureDaysToEnd = F, requireTimeAtRisk = T, minTimeAtRisk = 364,
  includeAllOutcomes = T, removePriorOutcome = T, recalibrate = T,
  calibrationPopulation = NULL)
}
\arguments{
\item{connectionDetails}{The connection details for extracting the data}

\item{cdmDatabaseSchema}{A string specifying the database containing the cdm}

\item{cohortDatabaseSchema}{A string specifying the database containing the target population cohort}

\item{outcomeDatabaseSchema}{A string specifying the database containing the outcome cohort}

\item{cohortTable}{A string specifying the table containing the target population cohort}

\item{outcomeTable}{A string specifying the table containing the outcome cohort}

\item{cohortId}{An iteger specifying the cohort id for the target population cohorts}

\item{outcomeId}{An iteger specifying the cohort id for the outcome cohorts}

\item{oracleTempSchema}{The temp schema require is using oracle}

\item{riskWindowStart}{The start of the period to predict the risk of the outcome occurring start relative to the target cohort start date}

\item{riskWindowEnd}{The end of the period to predict the risk of the outcome occurring start relative to the target cohort start date}

\item{addExposureDaysToEnd}{Add the riskWindowEnd to the cohort end rather than the cohort start to define the end of TAR}

\item{requireTimeAtRisk}{Require a minimum number of days observed in the time at risk period?}

\item{minTimeAtRisk}{If requireTimeAtRisk is true, the minimum number of days at risk}

\item{includeAllOutcomes}{Whether to include people with outcome who do not satify the minTimeAtRisk}

\item{removePriorOutcome}{Remove people with prior outcomes from the target population}

\item{recalibrate}{Recalibrate on new data}

\item{calibrationPopulation}{A data.frame of subjectId, cohortStartDate, indexes used to recalibrate the model on new data}
}
\value{
A list containing the model performance and the personal predictions for each subject in the target population
}
\description{
Apply the existing model Qstroke stroke risk using the standardised framework
}
\details{
This function applies Qstroke stroke risk to a target cohort and validates the performance
}
